update for mac

+++ SUB_TO_STOP.cpp
@@ -6,7 +6,7 @@
 int main(){
     zsock_t *sub = zsock_new_sub("tcp://172.20.10.3:6000", "CAR");
     char *topic;
-    int n;
+    uint64_t n;
     zsock_recv(sub, "s8",&topic, &n );
     while(n!= 1000000){
         zsock_recv(sub, "s8",&topic, &n );
